feat: moduralize utils functions
